+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/automation/singlesv/ReconcileSequencerLimitWithMemberTrafficTrigger.scala
@@ -11,7 +11,6 @@ import org.lfdecentralizedtrust.splice.automation.{
 }
 import org.lfdecentralizedtrust.splice.codegen.java.splice
 import org.lfdecentralizedtrust.splice.environment.SequencerAdminConnection
-import org.lfdecentralizedtrust.splice.sv.ExtraSynchronizerNode
 import org.lfdecentralizedtrust.splice.sv.store.SvDsoStore
 import org.lfdecentralizedtrust.splice.sv.util.SvUtil
 import org.lfdecentralizedtrust.splice.util.AssignedContract
@@ -35,7 +34,6 @@ class ReconcileSequencerLimitWithMemberTrafficTrigger(
     override protected val context: TriggerContext,
     store: SvDsoStore,
     sequencerAdminConnectionO: Option[SequencerAdminConnection],
-    extraSynchronizerNodes: Map[String, ExtraSynchronizerNode],
     trafficBalanceReconciliationDelay: NonNegativeFiniteDuration,
 )(implicit
     ec: ExecutionContext,
@@ -65,9 +63,7 @@ class ReconcileSequencerLimitWithMemberTrafficTrigger(
         memberId => {
           val synchronizerId = SynchronizerId.tryFromString(memberTraffic.payload.synchronizerId)
           val sequencerAdminConnection = SvUtil.getSequencerAdminConnection(
-            synchronizerId,
-            sequencerAdminConnectionO,
-            extraSynchronizerNodes,
+            sequencerAdminConnectionO
           )
           sequencerAdminConnection.getStatus
             .map(_.successOption.map(_.synchronizerId))