+++ apps/wallet/src/main/scala/org/lfdecentralizedtrust/splice/wallet/UserWalletManager.scala
@@ -60,7 +60,6 @@ class UserWalletManager(
     validatorTopupConfig: ValidatorTopupConfig,
     walletSweep: Map[String, WalletSweepConfig],
     autoAcceptTransfers: Map[String, AutoAcceptTransfersConfig],
-    supportsSoftDomainMigrationPoc: Boolean,
     dedupDuration: DedupDuration,
     txLogBackfillEnabled: Boolean,
     txLogBackfillingBatchSize: Int,
@@ -236,7 +235,6 @@ class UserWalletManager(
       // TODO(#12554): make it easier to configure the sweep functionality and guard better against operator errors (typos, etc.)
       walletSweep.get(endUserParty.toProtoPrimitive),
       autoAcceptTransfers.get(endUserParty.toProtoPrimitive),
-      supportsSoftDomainMigrationPoc,
       dedupDuration,
       txLogBackfillEnabled = txLogBackfillEnabled,
       txLogBackfillingBatchSize = txLogBackfillingBatchSize,