+++ cluster/pulumi/cluster/src/index.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { installNodePools } from './nodePools';
 import { installStorageClasses } from './storageClasses';