+++ cluster/pulumi/common-sv/src/synchronizer/cometbftConfig.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { Output } from '@pulumi/pulumi';
 import { config } from 'splice-pulumi-common/src/config';