+++ cluster/pulumi/common/src/auth0types.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 export interface Auth0ClientSecret {
   client_id: string;
   client_secret: string;