+++ cluster/pulumi/common/src/multiValidator.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 // Shared constants for multivalidator stuff across pulumi projects
 import { config } from './config';