+++ cluster/pulumi/operator/src/namespace.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { exactNamespace } from 'splice-pulumi-common';
 
 export const namespace = exactNamespace('operator');