+++ cluster/pulumi/operator/src/stacks/deployment.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { GitFluxRef } from 'splice-pulumi-common/src/operator/flux-source';
 import { createStackCR, EnvRefs } from 'splice-pulumi-common/src/operator/stack';