+++ cluster/pulumi/validator-runbook/src/utils.ts
@@ -1,3 +1,5 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { config } from 'splice-pulumi-common';
 
 export const VALIDATOR_NAMESPACE = config.optionalEnv('VALIDATOR_NAMESPACE') || 'validator';